test-lineage-BckEnd/server/requirements.txt
Flask
flask-cors
requests
PyYAML
python-dotenv
requests
psycopg2-binary